+++ conFusionServer/routes/leaderRouter.js
@@ -24,29 +24,47 @@ leaderRouter.route('/')
     }, (err) => next(err))
     .catch((err) => next(err));
 })
-.post(authenticate.verifyUser, (req, res, next) => {
-    Leaders.create(req.body)
-    .then((leader) => {
-        console.log('(Post) Leader created ', leader);
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(leader);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+.post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+    if (req.user.admin) {
+        Leaders.create(req.body)
+        .then((leader) => {
+            console.log('(Post) Leader created ', leader);
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(leader);
+        }, (err) => next(err))
+        .catch((err) => next(err));
+    }
+    else {
+        res.statusCode = 403;
+        res.end('(Post) You are not authorized to perform this operation!');    
+    }
 })
-.put(authenticate.verifyUser, (req, res, next) => {
-    res.statusCode = 403;
-    res.end('(Put) operation not supported on /leaders');
+.put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+    if (req.user.admin) {
+        res.statusCode = 403;
+        res.end('(Put) operation not supported on /leaders');
+    }
+    else {
+        res.statusCode = 403;
+        res.end('(Post) You are not authorized to perform this operation!');    
+    }
 })
-.delete(authenticate.verifyUser, (req, res, next) => {
-    Leaders.remove({})
-    .then((resp) => {
-        console.log('(Delete) Deleting all Leaders');
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(resp);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+.delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+    if (req.user.admin) {
+        Leaders.remove({})
+        .then((resp) => {
+            console.log('(Delete) Deleting all Leaders');
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(resp);
+        }, (err) => next(err))
+        .catch((err) => next(err));
+    }
+    else {
+        res.statusCode = 403;
+        res.end('(Post) You are not authorized to perform this operation!');    
+    }
 });
 
 
@@ -61,29 +79,47 @@ leaderRouter.route('/:leaderId/')
     }, (err) => next(err))
     .catch((err) => next(err));
 })
-.post(authenticate.verifyUser, (req, res, next)=> {
-    res.statusCode = 403;
-    res.end('(Post) operation not supported on /leaders/' + req.params.leaderId );   
+.post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next)=> {
+    if (req.user.admin) {
+        res.statusCode = 403;
+        res.end('(Post) operation not supported on /leaders/' + req.params.leaderId );       
+    }
+    else {
+        res.statusCode = 403;
+        res.end('(Post) You are not authorized to perform this operation!');    
+    }
 })
-.put(authenticate.verifyUser, (req, res, next) => {
-    Leaders.findByIdAndUpdate(req.params.leaderId, {
-        $set: req.body
-    }, {new : true})
-    .then((leader) => {
-        console.log('(Put) Will update leader ' + req.params.leaderId + ', for you!');
-        res = setRequest(res, 200);
-        res.json(leader);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+.put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+    if (req.user.admin) {
+        Leaders.findByIdAndUpdate(req.params.leaderId, {
+            $set: req.body
+        }, {new : true})
+        .then((leader) => {
+            console.log('(Put) Will update leader ' + req.params.leaderId + ', for you!');
+            res = setRequest(res, 200);
+            res.json(leader);
+        }, (err) => next(err))
+        .catch((err) => next(err));
+    }
+    else {
+        res.statusCode = 403;
+        res.end('(Put) You are not authorized to perform this operation!');    
+    }
 })
-.delete(authenticate.verifyUser, (req, res, next) => {
-    Leaders.findByIdAndRemove(req.params.leaderId)
-    .then((resp) => {
-        console.log('(Delete) Will delete leader ' + req.params.leaderId + '!');
-        res = setRequest(res, 200);
-        res.json(resp);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+.delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+    if (req.user.admin) {
+        Leaders.findByIdAndRemove(req.params.leaderId)
+        .then((resp) => {
+            console.log('(Delete) Will delete leader ' + req.params.leaderId + '!');
+            res = setRequest(res, 200);
+            res.json(resp);
+        }, (err) => next(err))
+        .catch((err) => next(err));
+    }
+    else {
+        res.statusCode = 403;
+        res.end('(Delete) You are not authorized to perform this operation!');    
+    }
 });
 
 module.exports = leaderRouter;
