+++ conFusionServer/routes/promoRouter.js
@@ -24,28 +24,46 @@ promoRouter.route('/')
     }, (err) => next(err))
     .catch((err) => next(err));
 })
-.post(authenticate.verifyUser, (req, res, next) => {
-    Promotions.create(req.body)
-    .then((promotion) => {
-        console.log('(Post) Promotion created ' + promotion);
-        res.statusCode = 200;
-        res.setHeader('Content-Type', 'application/json');
-        res.json(promotion);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+.post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+    if (req.user.admin) {
+        Promotions.create(req.body)
+        .then((promotion) => {
+            console.log('(Post) Promotion created ' + promotion);
+            res.statusCode = 200;
+            res.setHeader('Content-Type', 'application/json');
+            res.json(promotion);
+        }, (err) => next(err))
+        .catch((err) => next(err));
+    }
+    else {
+        res.statusCode = 403;
+        res.end('(Post) You are not authorized to perform this operation!'); 
+    }
 })
-.put(authenticate.verifyUser, (req, res, next) => {
-    res.statusCode = 403;
-    res.end('(Put) operation not supported on /promotions');
+.put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+    if (req.user.admin) {
+        res.statusCode = 403;
+        res.end('(Put) operation not supported on /promotions');
+    }
+    else {
+        res.statusCode = 403;
+        res.end('(Put) You are not authorized to perform this operation!'); 
+    }
 })
-.delete(authenticate.verifyUser, (req, res, next) => {
-    Promotions.remove({})
-    .then((resp) => {
-        console.log('(Delete) Deleting all the promotions');
-        res = setResponse(res, 200);
-        res.json(resp);
-    }, (err) => next(err))    
-    .catch((err) => next(err));
+.delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+    if (req.user.admin) {
+        Promotions.remove({})
+        .then((resp) => {
+            console.log('(Delete) Deleting all the promotions');
+            res = setResponse(res, 200);
+            res.json(resp);
+        }, (err) => next(err))    
+        .catch((err) => next(err));
+    }
+    else {
+        res.statusCode = 403;
+        res.end('(Delete) You are not authorized to perform this operation!'); 
+    }
 });
 
 
@@ -60,29 +78,47 @@ promoRouter.route('/:promotionId/')
     }, (err) => next(err))
     .catch((err) => next(err));
 })
-.post(authenticate.verifyUser, (req, res, next) => {
-    res.statusCode = 403;
-    res.end('(Post) operation not supported on /promotions/' + req.params.promotionId);
+.post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+    if (req.user.admin) {
+        res.statusCode = 403;
+        res.end('(Post) operation not supported on /promotions/' + req.params.promotionId);
+    }
+    else {
+        res.statusCode = 403;
+        res.end('(Post) You are not authorized to perform this operation!'); 
+    }
 })
-.put(authenticate.verifyUser, (req, res, next) => {
-    Promotions.findByIdAndUpdate(req.params.promotionId, {
-        $set: req.body
-    }, {new : true})
-    .then((promotion) => {
-        console.log('(Put) Updating promotion ' + req.params.promotionId + ', for you!');
-        res = setResponse(res, 200);
-        res.json(promotion);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+.put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+    if (req.user.admin) {
+        Promotions.findByIdAndUpdate(req.params.promotionId, {
+            $set: req.body
+        }, {new : true})
+        .then((promotion) => {
+            console.log('(Put) Updating promotion ' + req.params.promotionId + ', for you!');
+            res = setResponse(res, 200);
+            res.json(promotion);
+        }, (err) => next(err))
+        .catch((err) => next(err));
+    }
+    else {
+        res.statusCode = 403;
+        res.end('(Put) You are not authorized to perform this operation!'); 
+    }
 })
-.delete(authenticate.verifyUser, (req, res, next) => {
-    Promotions.findByIdAndDelete(req.params.promotionId)
-    .then((resp) => {
-        console.log('(Delete) Will delete promotion ' + req.params.promotionId + '!');
-        res = setResponse(res, 200);
-        res.json(resp);
-    }, (err) => next(err))
-    .catch((err) => next(err));
+.delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
+    if (req.user.admin) {
+        Promotions.findByIdAndDelete(req.params.promotionId)
+        .then((resp) => {
+            console.log('(Delete) Will delete promotion ' + req.params.promotionId + '!');
+            res = setResponse(res, 200);
+            res.json(resp);
+        }, (err) => next(err))
+        .catch((err) => next(err));
+    }
+    else {
+        res.statusCode = 403;
+        res.end('(Delete) You are not authorized to perform this operation!'); 
+    }
 });
 
 module.exports = promoRouter;
